feat: 优化编译和波形显示 - 清理 clippy 警告（needless_range_loop, format_in_format_args） - 波形 1S 宽度从 200 像素缩小到 60 像素 - 后端 target_pixels 同步调整

+++ src/alignment_v3.rs
@@ -394,7 +394,7 @@ pub fn find_all_alignments_v3(
     let mut deduped = Vec::new();
     for r in results {
         let too_close = deduped.iter().any(|existing: &AlignmentResult| {
-            ((r.offset_samples as isize - existing.offset_samples as isize).unsigned_abs() as usize) < min_gap
+            (r.offset_samples as isize - existing.offset_samples as isize).unsigned_abs() < min_gap
         });
         if !too_close {
             deduped.push(r);
